Enhance bulk upload, add monthly filtering, and improve student registration views
